refactor: use new_createPages in examples and e2e tests
