+++ examples/42_react-tweet/src/entries.tsx
@@ -1,4 +1,4 @@
-import { createPages } from 'waku';
+import { new_createPages as createPages } from 'waku/router/server';
 
 import { RootLayout } from './templates/root-layout';
 import { HomePage } from './templates/home-page';
